refactor: arrange AgentStatus
